Create README.md
This system is designed to control room environment within a small office. The key interests in this application are control of lighting, heating, energy saving measures, and user feedback displays.

The system was designed on two Arduino Uno microcontrollers. One microcontroller was given a set of sensors (a temperature reader, a PIR sensor, and a reed switch), this microcontroller would take readings from these sensors and send them via Bluetooth to a receiving microcontroller. The receiving microcontroller was given a set of output systems to control room environment; a relay was used to control lighting (for the project the power cable of a lamp was placed through the relay to enact this function) and a stepper motor was used to control heating (I have included an image in this repository to show a basic example of how this function could be completed). The receiving microcontroller was also given a RTC for time-based functions.

Most operations were implemented on the receiving microcontroller, energy saving functions were implemented to; reduce energy wasted on heating during inactivity within the room, reduce energy wasted on heating during overcrowding within the room, prevent energy being used on heating or lighting outside office hours. Feedback displays were implemented on this microcontroller as well and were intended to be used with a program like Tera Term to show outputs.

Images of design plan and finished product have been included in this repository to help explain system further.